rename files + fix CSS convention
