+++ src/components/WalletInfo/WalletInfo.scss
@@ -1,10 +1,10 @@
 .wallet-info {
-  &--menu {
+  &__menu {
     background-color: var(--c-grey);
     border: .06rem solid var(--c-dark-border);
     border-radius: .3rem;
   }
-  &--banner {
+  &__banner {
     background-color: transparent;
   }
   padding: 1rem!important;
