+++ src/compositions/TopBar/TopBar.scss
@@ -9,6 +9,7 @@ $component: top-bar;
     width: 100%;
     height: 3rem;
     padding: 0.5rem;
+    background: var(--c-grey);
     position: relative;
     z-index: 1;
 
@@ -32,6 +33,7 @@ $component: top-bar;
         background: transparent;
         animation: mobile-in ease-in-out 0.5s forwards;
     }
+    // have to find a way to activate this keyframes on click of the close icons but not in the initialization of the component
     &__desktop {
         background: var(--c-grey);
         animation: mobile-out ease-in-out 0.5s forwards;
@@ -63,7 +65,6 @@ $component: top-bar;
     }
 }
 
-// have to find a way to activate this keyframes on click of the close icons
 @keyframes mobile-out {
     from {
         background: transparent;
